added the combine reducer and created externalData in its own file
